Check for existing Chocolatey installation in setup.ps1

+++ setup.ps1
@@ -126,11 +126,15 @@ if ((Test-Path -Path $PROFILE) -and (winget list --name "OhMyPosh" -e) -and ($fo
 }
 
 # Choco install
-try {
-    Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))
-}
-catch {
-    Write-Error "Failed to install Chocolatey. Error: $_"
+if (!(Test-Path -Path "C:\ProgramData\chocolatey\bin\choco.exe")) {
+    try {
+        Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))
+    }
+    catch {
+        Write-Error "Failed to install Chocolatey. Error: $_"
+    }
+} else {
+    Write-Host "Chocolatey is already installed. Skipping installation."
 }
 
 # Terminal Icons Install
